refactor: :recycle: Refactor some code and if the .env file does not exists then it will be created
